Product Admin - Secure Admin & Profile Views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404, redirect, render, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 # Q from Django to handle filtering search by name and or description
@@ -64,6 +65,7 @@ def all_products(request):
     
     return render(request, 'products/products.html', context)
 
+
 def product_detail(request, product_id):
     """ A view to show individual product details """
     
@@ -76,8 +78,15 @@ def product_detail(request, product_id):
     
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """ Add a product to the database """
+    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, that is a job for Admin!')
+        return redirect(reverse('home'))
+
     
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -97,8 +106,14 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """ Edit an existing product in the database """
+    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, that is a job for Admin!')
+        return redirect(reverse('home'))
     
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -121,8 +136,14 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the database """
+    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, that is a job for Admin!')
+        return redirect(reverse('home'))
     
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
